refactor : 팝업스토어 테스트코드 수정
